+++ NEWS.md
@@ -11,6 +11,8 @@ None
 `hai_auto_c50()`
 4. Fix #246 - Add boilerplate functions for glmnet `hai_glmnet_data_prepper()` and
 `hai_auto_glmnet()`
+5. Fix #249 - Add boilerplate functions for cubist `hai_cubist_data_prepper()` and
+`hai_auto_cubist()`
 
 ## Minor Fixes and Improvements
 1. Fix #240 = Add parameter arguments as attributes to list output for boilerplate